+++ src/app/src/screens/Pod/Navigation.ts
@@ -1,0 +1,5 @@
+import { pathTo } from "../../Navigator";
+
+export const getPodUrl = (id: string, view: string) => {
+  return pathTo(`/screens/pod/${encodeURIComponent(id)}/${encodeURIComponent(view)}`);
+};